backend/requirements.txt
Flask==3.0.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0
python-dotenv==1.0.1
Flask-Limiter>=3.0.0
gunicorn==22.0.0
